Add files via upload
Credit card fraud detection refers to the set of policies, tools, methodologies, and practices that credit card companies and financial institutions use to prevent fraudulent purchases, both online and in-store. It involves using various techniques and technologies to identify potentially fraudulent transactions in real-time or post-transaction analysis. The goal is to minimize financial losses for both cardholders and card issuers by quickly identifying and stopping unauthorized or suspicious transactions.